% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{copuladata}
\alias{copuladata}
\title{Simulated copula dataset}
\format{An object of class \code{matrix}.}
\usage{
data(copuladata)
}
\description{
Simulated bivariate dataset with a gaussian component
and an exponential one, with a certain dependency structure.
}
\examples{
data(copuladata)
}
\keyword{datasets}
