% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{copulahmmdata}
\alias{copulahmmdata}
\title{Simulated copula dataset with underlying Markov Model}
\format{An object of class \code{matrix}.}
\usage{
data(copulahmmdata)
}
\description{
Simulated bivariate dataset with a gaussian component
and an exponential one, with a certain dependency structure
and an underlying 2-state Markov Model.
}
\examples{
data(copulahmmdata)
}
\keyword{datasets}
