% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardbackward.R
\name{forwardbackward}
\alias{forwardbackward}
\title{Forward-Backward Function}
\usage{
forwardbackward(nStates, nu, A, B, bT)
}
\arguments{
\item{nStates}{an integer representing the number of state of the Hidden Markov Model.}

\item{nu}{a vector of the initial probabilities.}

\item{A}{the transition matrix.}

\item{B}{a matrix containing the values of the emission probabilities}

\item{bT}{the vector of the beginning times for the statistical units}
}
\value{
The function returns a numeric value indicating the distance between the two curves.
}
\description{
This function allows you to perform the forward-backward algorithm for a Hidden Markov Model
}
\references{
Rabiner, L. R. (1989). A tutorial on hidden Markov models and selected applications in speech recognition. Proceedings of the IEEE, 77(2), 257-286.
}
\keyword{distance}
