% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{simulatedFD}
\alias{simulatedFD}
\title{Simulated multivariate functional dataset with underlying Markov Model}
\format{An object of class \code{funData}.}
\usage{
data(simulatedFD)
}
\description{
Simulated bivariate functional dataset with
an underlying 2-state Markov Model.
}
\examples{
data(simulatedFD)
}
\keyword{datasets}
